refactor code; add header files
