+++ src/classes/Home/home.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include<string>
+
+#include "../Config/config.hpp"
+#include "../Category/category.hpp"
+
+void buildHome(Config *conf, Category *categories[], int categories_length, std::string path);
