+++ src/classes/Page/head.hpp
@@ -1,0 +1,5 @@
+#pragma once
+#include<string>
+#include"../Config/config.hpp"
+
+std::string html_head(Config* conf, std::string text);
