+++ src/helpers/helpers.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <string>
+
+char getLower(char c);
+std::string toLowerCase(std::string text);
+void printError(int linenumber,std::string text);
