+++ src/helpers/linkparser.cpp
@@ -1,0 +1,155 @@
+#include <string>
+#include "helpers.hpp"
+
+using namespace std;
+
+//  [urlText](url) ->  <a href='url'>urlText</a>
+// ![altText](image) -> <img src='' alt=''>
+string static parseLinks(string text, string path)
+{
+    string label = text;
+    unsigned int index = 0;
+    string newText = "";
+    string urlText = "";
+    string url = "";
+    bool isImage = false;
+    bool isTag = false;
+
+    for (; index < text.size();)
+    {
+
+        // If tag
+        if (isTag == true)
+        {
+            string tag = "";
+            while (text[index] != '}' && index < text.size())
+            {
+                tag += text[index];
+                index += 1;
+            }
+            index += 2; // Increment index to new text;
+
+            string filename = toLowerCase(tag);
+            string filepath = path + filename + ".html";
+            newText += "<a class='tag' href='" + filepath + "'>{{" + tag + "}}</a>";
+            isTag = false;
+            continue;
+        }
+        // Check if tag
+        if (text[index] == '{' && text[index + 1] == '{')
+        {
+            isTag = true;
+            index += 2;
+            continue;
+        }
+        // Enter into linking
+        // ![ -> image flag true
+        if (text[index] == '!' && text[index + 1] == '[')
+        {
+            isImage = true;
+            index++;
+            continue;
+
+            //
+        }
+        else if (text[index] == '[')
+        {
+
+            // Crude implementation for images
+            if (index >= 1 && text[index - 1] == '!')
+            {
+                isImage = true;
+            }
+            else
+            {
+                isImage = false;
+            }
+
+            index++;
+            // Loop through urltext
+            while (index < text.size() && text[index] != ']')
+            {
+
+                // Broken [text [text](link) case
+                if (text[index] == '[')
+                {
+                    newText += '[' + urlText;
+                    urlText = "";
+                    index++;
+                }
+
+                urlText += text[index];
+                index++;
+            }
+
+            // EndofUrlText
+            // [urltext]
+            if (text[index] == ']' && index < text.size())
+            {
+                index++;
+                if (text[index] == '(' && index < text.size())
+                {
+                    index++;
+                    // Continue
+                    while (index < text.size() && text[index] != ')')
+                    {
+
+                        // Error [link](abc.com [link](abc.com)
+                        if (text[index] == '[')
+                        {
+                            newText += '[' + urlText + "](" + url;
+                            urlText = "";
+                            break;
+                        }
+                        url += text[index];
+                        index++;
+                    }
+
+                    // Complete parsing
+                    if (text[index] == ')' && index < text.size())
+                    {
+
+                        if (isImage)
+                        {
+                            index++;
+                            newText += "<img src='" + url + "' alt='" + urlText + "'>";
+                        }
+                        else
+                        {
+                            index++;
+                            newText += " <a href='" + url + "'>" + urlText + "</a>";
+                        }
+                        url = "";
+                        urlText = "";
+                        isImage = false;
+                        continue;
+                    }
+
+                    if (index == text.size())
+                    {
+                        newText += '[' + urlText + "](" + url;
+                        continue;
+                    }
+                }
+                else
+                {
+                    newText += '[' + urlText + ']';
+                    urlText = "";
+                    continue;
+                }
+            }
+            else
+            {
+                newText += '[' + urlText;
+                urlText = "";
+                continue;
+            }
+        }
+        else
+        {
+            newText += text[index];
+            index++;
+        }
+    }
+    return newText;
+}
