+++ src/main.cpp
@@ -2,7 +2,6 @@
 #include"mizi.cpp"
 using namespace std;
 
-
 int main(int argc,char *argv[]){
     
     for (int i=0 ; i < argc; i++){
